requirements.txt
opencv_contrib_python==4.5.1.48
numpy==1.19.4
Keras==2.4.3
eventhandler==1.2.1
requests==2.25.1
Flask==1.1.2
gpiozero==1.6.2
picamera==1.13
Pillow==8.2.0
scikit_learn==0.24.1
